+++ presentation/src/main/java/com/weit2nd/presentation/ui/mypage/MyPageState.kt
@@ -3,7 +3,7 @@ package com.weit2nd.presentation.ui.mypage
 data class MyPageState(
     val profileImage: String? = null,
     val nickname: String = "",
-    val coin: Long = 0,
+    val coin: Int = 0,
     val isLogoutDialogShown: Boolean = false,
     val isWithdrawDialogShown: Boolean = false,
 )
